+++ SingleTraceSubprocessor.h
@@ -7,8 +7,8 @@ class SingleTraceSubprocessor : public SingleChannelSubprocessor
   SingleTraceSubprocessor(std::string name,
 		     module_index_t idx,
 		     int nbins,
-		     int upperLimit, 
-		     int lowerLimit=0);
+		     double upperLimit, 
+		     double lowerLimit=0);
   virtual void processSubevent(eventinfo_t ei);
 };
 #endif
